requirements.txt
anyio==4.5.2
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
typing-extensions==4.13.2
urllib3==2.2.3
